refactor(vsmgr: poster): rename new & old impls
