+++ venus-sector-manager/modules/poster/poster_test.go
